# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the core API for Sispeg Project in Azmeela",
        "title": "Sispeg API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/employees": {
            "get": {
                "description": "Get a list of all employees",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Get all employees",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.Employee"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new employee",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Create an employee",
                "parameters": [
                    {
                        "description": "Employee object",
                        "name": "employee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.Employee"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.Employee"
                        }
                    }
                }
            }
        },
        "/employees/{id}": {
            "get": {
                "description": "Get an employee by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Get a single employee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.Employee"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an employee profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Update an employee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Employee object",
                        "name": "employee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.Employee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.Employee"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an employee profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Delete an employee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/overtimes": {
            "get": {
                "description": "Get a list of overtime requests",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "overtimes"
                ],
                "summary": "Get overtimes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.Overtime"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Request a new overtime",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "overtimes"
                ],
                "summary": "Request overtime",
                "parameters": [
                    {
                        "description": "Overtime request details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.OvertimeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.Overtime"
                        }
                    }
                }
            }
        },
        "/overtimes/{id}/status": {
            "put": {
                "description": "Approve or reject an overtime request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "overtimes"
                ],
                "summary": "Update overtime status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Overtime ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.OvertimeStatusUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_azmeela_sispeg-api_internal_domain.Employee": {
            "type": "object",
            "properties": {
                "base_salary": {
                    "type": "integer"
                },
                "bio": {
                    "type": "string"
                },
                "contract_end": {
                    "type": "string"
                },
                "contract_start": {
                    "type": "string"
                },
                "cv": {
                    "type": "string"
                },
                "date_sign": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "type_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "github_com_azmeela_sispeg-api_internal_domain.Overtime": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "employee_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "description": "0: created, 1: approved, 2: disapproved",
                    "type": "integer"
                },
                "time_in": {
                    "type": "string"
                },
                "time_out": {
                    "type": "string"
                }
            }
        },
        "github_com_azmeela_sispeg-api_internal_domain.OvertimeRequest": {
            "type": "object",
            "required": [
                "date",
                "description",
                "time_in",
                "time_out"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "time_in": {
                    "type": "string"
                },
                "time_out": {
                    "type": "string"
                }
            }
        },
        "github_com_azmeela_sispeg-api_internal_domain.OvertimeStatusUpdate": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2
                    ]
                }
            }
        },
...
    }
}